pyproject.toml
[tool.poetry]
name = "money"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
django-rest-framework = "^0.1.0"
pytest-django = "^4.8.0"
dj-database-url = "^2.1.0"
mysqlclient = "^2.2.4"
django-filter = "^24.2"


[tool.poetry.group.dev.dependencies]
factory-boy = "^3.3.0"
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "money",
]
DJANGO_SETTINGS_MODULE = "money.settings"
python_files = [
    "tests.py",
    "test_*.py",
]
